# docs/swagger.yaml
basePath: /
definitions:
  dto.CreateAssetRequest:
    properties:
      currency:
        maxLength: 10
        minLength: 2
        type: string
      exchange:
        maxLength: 100
        minLength: 2
        type: string
      external_platform_id:
        maxLength: 100
        minLength: 2
        type: string
      instrument_type:
        type: string
      isin:
        maxLength: 50
        minLength: 10
        type: string
      name:
        maxLength: 200
        minLength: 1
        type: string
      symbol:
        maxLength: 20
        minLength: 1
        type: string
    required:
    - exchange
    - instrument_type
    - isin
    - name
    - symbol
    type: object
  dto.CreateTransactionRequest:
    properties:
      price:
        type: number
      quantity:
        type: number
      txn_date:
        type: string
      txn_type:
        type: string
      user_asset_id:
        type: integer
    required:
    - price
    - quantity
    - txn_date
    - txn_type
    - user_asset_id
    type: object
...
  dto.UpdateAssetRequest:
    properties:
      currency:
        maxLength: 10
        minLength: 2
        type: string
      exchange:
        maxLength: 100
        minLength: 2
        type: string
      external_platform_id:
        maxLength: 100
        minLength: 2
        type: string
      instrument_type:
        type: string
      isin:
        maxLength: 50
        minLength: 10
        type: string
      name:
        maxLength: 200
        minLength: 1
        type: string
      symbol:
        maxLength: 20
        minLength: 1
        type: string
    type: object
  dto.UpdateTransactionRequest:
    properties:
      price:
        type: number
      quantity:
        type: number
      txn_date:
        type: string
      txn_type:
        type: string
    type: object
  model.Asset:
    properties:
      created_at:
        type: string
      currency:
        type: string
      exchange:
        type: string
      external_platform_id:
        type: string
      id:
        type: integer
      instrument_type:
        type: string
      isin:
        type: string
      name:
        type: string
      symbol:
        type: string
    type: object
  model.Transaction:
    properties:
      created_at:
        type: string
      id:
        type: integer
      price:
        type: number
      quantity:
        type: number
      txn_date:
        type: string
      txn_type:
        type: string
      user_asset_id:
        type: integer
    type: object
...
  util.ErrorBody:
    properties:
      error: {}
    type: object
info:
  contact: {}
  description: API for managing users and assets in MF Stock Tracker.
  title: MF Stock Tracker API
  version: "1.0"
paths:
  /api/assets/:
    get:
      description: Retrieve assets with pagination
      parameters:
      - description: 'Number of records to return (default: 50, max: 200)'
        in: query
        name: limit
        type: integer
      - description: 'Number of records to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/model.Asset'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorBody'
      security:
      - BearerAuth: []
      summary: List assets
      tags:
      - assets
    post:
      consumes:
      - application/json
      description: Create a new asset
      parameters:
      - description: Create asset payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAssetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Asset'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create asset
      tags:
      - assets
  /api/assets/{assetId}:
    delete:
      description: Delete an asset by ID
      parameters:
      - description: Asset ID
        format: int64
        in: path
        name: assetId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorBody'
      security:
      - BearerAuth: []
      summary: Delete asset
      tags:
      - assets
    get:
      description: Retrieve a single asset by its ID
      parameters:
      - description: Asset ID
        format: int64
        in: path
        name: assetId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Asset'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get asset by ID
      tags:
      - assets
    put:
      consumes:
      - application/json
      description: Update an existing asset by ID
      parameters:
      - description: Asset ID
        format: int64
        in: path
        name: assetId
        required: true
        type: integer
      - description: Update asset payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAssetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update asset
      tags:
      - assets
  /api/transactions:
    get:
      description: Get all transactions across all user assets for a given user with
        pagination
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: 'Number of records to return (default: 50, max: 200)'
        in: query
        name: limit
        type: integer
      - description: 'Number of records to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Transaction'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorBody'
      security:
      - BearerAuth: []
      summary: List transactions for a user
      tags:
      - transactions
    post:
      consumes:
      - application/json
      description: Create a new transaction for a user asset
      parameters:
      - description: Create transaction payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create transaction
      tags:
      - transactions
  /api/transactions/{txnId}:
    delete:
      description: Delete a transaction by its ID
      parameters:
      - description: Transaction ID
        in: path
        name: txnId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorBody'
      security:
      - BearerAuth: []
      summary: Delete transaction
      tags:
      - transactions
    put:
      consumes:
      - application/json
      description: Update an existing transaction by ID
      parameters:
      - description: Transaction ID
        in: path
        name: txnId
        required: true
        type: integer
      - description: Update transaction payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update transaction
      tags:
      - transactions
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"